requirements.txt
pymongo
flask
bs4
python-dotenv